+++ contributions/catalog/controllers/contribute.py
@@ -9,6 +9,7 @@ from flask import (
 
 from controllers.config import Config
 from models.contribution_utilities import to_contribution
+from utils import jsonutil
 
 bp = Blueprint('contribute', __name__, url_prefix='/contribute')
 
@@ -43,9 +44,11 @@ def create():
         # result = dict((key, request.form.getlist(key) if len(request.form.getlist(key)) > 1 else request.form.getlist(key)[0]) for key in request.form.keys())
 
         contribution = to_contribution(result)
-        # print(contribution)
+
+        # add contributionAdmins to the json_contiubtion
+        contribution = jsonutil.add_contribution_admins(contribution)
+
         json_contribution = json.dumps(contribution, indent=4)
-        print(json_contribution)
         response, s = post(json_contribution)
         if response:
             return render_template('contribute/submitted.html', user=session["name"])
@@ -72,7 +75,7 @@ def search_results(search):
 def post(json_data):
     headers = {
         'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + Config.AUTHENTICATION_TOKEN
+        'Authorization': 'Bearer ' + session['oauth_token']['access_token']
     }
     try:
         # Setting up post request